src/main/kotlin/com/study/kotlinboard/service/PostService.kt
package com.study.kotlinboard.service

import com.study.kotlinboard.exception.PostNotDeletableException
import com.study.kotlinboard.exception.PostNotFoundException
import com.study.kotlinboard.repository.PostRepository
import com.study.kotlinboard.service.dto.PostCreateRequestDto
import com.study.kotlinboard.service.dto.PostUpdateRequestDto
import com.study.kotlinboard.service.dto.toEntity
import org.springframework.data.repository.findByIdOrNull
import org.springframework.stereotype.Service
import org.springframework.transaction.annotation.Transactional

@Service
@Transactional(readOnly = true)
class PostService(
    private val postRepository: PostRepository,
) {
    @Transactional
    fun createPost(postCreateRequestDto: PostCreateRequestDto): Long {
        return postRepository.save(postCreateRequestDto.toEntity()).id
    }

    @Transactional
    fun updatePost(id: Long, postUpdateRequestDto: PostUpdateRequestDto): Long {
        val post = postRepository.findByIdOrNull(id) ?: throw PostNotFoundException()
        post.update(postUpdateRequestDto)
        return id
    }

    @Transactional
    fun deletePost(id: Long, deletedBy: String): Long {
        val post = postRepository.findByIdOrNull(id) ?: throw PostNotFoundException()
        if (post.createdBy != deletedBy) throw PostNotDeletableException()
        postRepository.delete(post)
        return id
    }
}
